Fixing Condition Chances.
Previously conditions had a Double value but the inputs used were of type Int. Given the code I believe it mad more sense to change to change how to type Int.

+++ All_Moves.cpp
@@ -1,7 +1,7 @@
 #include <string>
 #include <map>
 #include "Move.h"
-std::map<std::string, double> emptyMap;
+std::map<std::string, int> emptyMap;
 
 
 std::map<std::string, Move> moveMap = {
